% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regs.R
\name{regs_context}
\alias{regs_context}
\title{Generate LLM-friendly context for the api.regulations.gov package}
\usage{
regs_context()
}
\value{
Character string (invisibly), also printed
}
\description{
Generate LLM-friendly context for the api.regulations.gov package
}
